Move SyncClient.ts to utils/
